+++ tablesalt/lib/tablesalt/rspec/custom_matchers.rb
@@ -1,4 +1,6 @@
 # frozen_string_literal: true
 
+require_relative "custom_matchers/define_thread_accessor"
 require_relative "custom_matchers/define_thread_reader"
+require_relative "custom_matchers/define_thread_writer"
 require_relative "custom_matchers/isolate"