+++ tablesalt/lib/tablesalt/rspec/shared_examples.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "shared_examples/a_thread_accessor"
-require_relative "shared_examples/a_thread_reader"
-require_relative "shared_examples/a_thread_writer"