+++ tablesalt/lib/tablesalt/spec_helper.rb
@@ -2,4 +2,3 @@
 
 require_relative "rspec/custom_matchers"
 require_relative "rspec/config"
-require_relative "rspec/shared_examples"